Adding confirm prompt for project rm

+++ src/commands/project-rm.js
@@ -1,13 +1,32 @@
+import prompt from 'prompt';
+import { getPrompt } from '../libs/prompt';
 import chalk from 'chalk';
 import Table from 'cli-table';
 
 import config from '../config';
 import * as specActions from '../actions/spec-actions';
 
+const prompts = [{
+  name: 'confirm',
+  description: 'Are you sure you\'d like to remove this project? [Y/n]'
+}];
+
 export default async function({getState, dispatch}, uuid) {
-	console.log(chalk.gray(`Removing project with id '${uuid}'`));
 
-	await dispatch(specActions.remove(uuid));
+	prompt.start();
 
-	console.log(chalk.red(`Project ${uuid} removed`));
+	prompt.message = '';
+	prompt.delimiter = '';
+
+	const { confirm } = await getPrompt(prompts);
+
+	if (confirm == 'Yes' || confirm == 'Y') {
+
+		console.log(chalk.gray(`Removing project with id '${uuid}'`));
+
+		await dispatch(specActions.remove(uuid));
+
+		console.log(chalk.red(`Project '${uuid}' removed`));
+
+	}
 }
